+++ src/Media.java
@@ -1,29 +1,26 @@
 
 public abstract class Media {
-	private int id;
+	private long id;
 	private CheckoutCard card;
 	private Type type;
-	
-	public Media(int id, Type type) {
+
+	public Media(long id, Type type) {
 		this.id = id;
 		this.type = type;
 	}
-	
-	public int getId() {
+
+	public long getId() {
 		return id;
 	}
-	
-	public void setId(int id) {
+
+	public void setId(long id) {
 		this.id = id;
 	}
-	
+
 	public Type getType() {
 		return type;
 	}
-	
-//	public void accept(Visitor visitor) {
-//		visitor.visit(this);
-//	}
+
 	public static enum Type {
 		Book
 	}
